Updates to data description, some misc formatting stuff.
Note: I have to keep taking out the fig=TRUE or my doc won't compile. So far that's the only thing that seems to throw the error.

Fixed up the variables table (still more to do). Changed "factor" back to "binary" since I didn't want to lose that info. I think maybe this would be better as two tables so I'll work on that. I'd like to have a record of data transformations but I like the idea of putting "factor" so we know what functions will work how.

Also worked on interpreting some of the variables and thinking about predictors vs. responses.

Changed the headings to subsections and had to get rid of some \newline commands (they were causing the doc to not compile).

Added a section at the end for scatterplots but this is mostly to mess around with the data and look at it. Incomplete.