Updated figure in final section
I figured out captions! See my code for details.

Also I have to comment out the \SweaveOpts thing for it to run but I guess until we solve it we'll have to go back and forth on this. I'm guessing one of us is missing something on our computer which, once installed, will solve the problem.

	modified:   writeup.Rnw